backend/requirements.txt
faiss-cpu==1.7.4
openai==1.3.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
sqlite3==3.42.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
typing==3.7.4.3
pathlib==1.0.1